Add namespace for client employment journey
