+++ config/routes.rb
@@ -155,7 +155,9 @@ Rails.application.routes.draw do
 
       namespace :income, constraints: -> (_) { FeatureFlags.means_journey.enabled? } do
         edit_step :what_is_clients_employment_status, alias: :employment_status
-        crud_step :employer_details, alias: :employer_details, param: :employment_id
+        namespace :client do
+          crud_step :employer_details, alias: :employer_details, param: :employment_id
+        end
         show_step :employed_exit
         edit_step :did_client_lose_job_being_in_custody, alias: :lost_job_in_custody
         edit_step :current_income_before_tax, alias: :income_before_tax
